% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_halton.R
\name{make_draws}
\alias{make_draws}
\title{Generate Various Types of Random and Quasi-Random Draws}
\usage{
make_draws(n_draws, n_dim, type, ...)
}
\arguments{
\item{n_draws}{The number of draws (or points) to generate.}

\item{n_dim}{The number of dimensions for each draw.}

\item{type}{A character string specifying the type of draws. Options are:
\itemize{
  \item `"standard-halton"`: Standard, unscrambled Halton sequence.
  \item `"scrambled-halton-sqrt"`: Halton sequence scrambled with the square-root method.
  \item `"scrambled-halton-neg-sqrt"`: Halton sequence scrambled with the negative square-root method.
  \item `"scrambled-halton-rand-mult"`: Halton sequence scrambled with a random multiplier.
  \item `"scrambled-halton-rand-perm"`: Halton sequence scrambled with a random permutation.
  \item `"scrambled-halton-atanassov"`: Halton sequence scrambled with Atanassov's method.
  \item `"standard-sobol"`: Standard Sobol sequence.
  \item `"scrambled-sobol"`: Scrambled Sobol sequence.
  \item `"pseudo-random"`: Standard pseudo-random draws from a uniform distribution.
}}

\item{...}{Additional arguments passed to specific draw functions (e.g., `seed`
for scramblers that use randomization).}
}
\value{
A matrix of draws with `n_draws` rows and `n_dim` columns.
}
\description{
A unified interface for creating different types of draws for simulation,
including standard and scrambled Halton sequences based on the methods
described by Kolenikov (2012).
}
\examples{
# Generate 100 draws in 6 dimensions (5 & 6 are correlated)

# Standard Halton sequence
draws1 <- make_draws(100, 6, type = "standard-halton")

# Scrambled Halton sequence using the square-root method
draws2 <- make_draws(100, 6, type = "scrambled-halton-neg-sqrt")
#'
# Scrambled Halton sequence with a random permutation and a seed
draws3 <- make_draws(100, 6, type = "scrambled-halton-rand-perm", seed = 123)

# Scrambled Halton sequence with atanassov
draws4 <- make_draws(100, 6, type = "scrambled-halton-atanassov")

# Plot to compare
par(mfrow = c(1, 4))
plot(draws1[,5:6], main = "Standard Halton")
plot(draws2[,5:6], main = "Scrambled (Neg-Sqrt)")
plot(draws3[,5:6], main = "Scrambled (Rand Perm)")
plot(draws4[,5:6], main = "Scrambled (Atanassov)")

}
